Prevent unneeded join requests & added some get/set functions
If the module remains powered, the joined status values remain in tact.
This allows to continue using the same session and just continue.
This is possible using either OTAA or ABP.
So the default assumption is now that there is a "joined" state, and only perform a true join when the result of a TX attempt returns "not_joined".

This does save a number of unneeded join requests and thus the number of downlink packets.

In order to keep using an ABP session, one must call `setLastUsedJoinMode` before calling an init function. 
Right after calling the constructor. (so it should be part of the constructor?)

For ABP there is also a function added to set the frame counters, in order to continue the session.

The internal status is now also monitored, but it is not yet clear if that's useful right now.
The "joined" flag for instance is never set to '1'.